[client] refactoring wip
