+++ client/src/components/index.ts
@@ -1,0 +1,5 @@
+export * from './Jingle';
+export * from './Now';
+export * from './Player';
+export * from './SongModal';
+export * from './UserModal';
