+++ client/src/stores/index.tsx
@@ -1,0 +1,4 @@
+export * from './CommonStore';
+export * from './MainStore';
+export * from './SongStore';
+export * from './UserStore';
